feat: Dominio Estado

+++ algaworks-spring-specialist/src/main/java/com/algaworks/algafood/infrastructure/repository/EstadoRepositoryImpl.java
@@ -4,6 +4,7 @@ import com.algaworks.algafood.domain.model.Estado;
 import com.algaworks.algafood.domain.repository.EstadoRepository;
 import jakarta.persistence.EntityManager;
 import lombok.AllArgsConstructor;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -34,8 +35,13 @@ public class EstadoRepositoryImpl implements EstadoRepository {
 
         @Override
         @Transactional
-        public void remover(Estado estado) {
-            estado = buscar(estado.getId());
+        public void remover(Long estadoId) {
+            Estado estado = buscar(estadoId);
+
+            if (estado == null) {
+                throw new EmptyResultDataAccessException(1);
+            }
+
             entityManager.remove(estado);
         }
